changes
extended the simulator to include other params like memory and so on

+++ Assets/Scripts/Core/Data/EvaluationData.cs
@@ -11,7 +11,7 @@ public class EvaluationData
     // Stats
     public float TimeTaken;
     public float MeasuredTimeMs;
-    public long MemoryUsedBytes;
+    public float MemoryUsedBytes;
 
     public string Message;
 }
